Added new Features made the game dynamic 😘
